implemented react router and works, still needs customization
