+++ package.json
@@ -27,6 +27,7 @@
   "dependencies": {
     "axios": "^0.15.3",
     "react": "^15.4.2",
-    "react-dom": "^15.4.2"
+    "react-dom": "^15.4.2",
+    "react-router": "^3.0.2"
   }
 }
