--- public/components/city.jsx
@@ -1,2 +1,0 @@
-import React from 'react';
-import WeatherCard from './weather-card';
