+++ public/components/home.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+class Home extends React.Component {
+  constructor(props) {
+    super(props);
+  }
+
+  render() {
+    return (
+      <form>
+        <input type="text" placeholder="Search City" />
+        <button>Search</button>
+      </form>
+    );
+  }
+}
+
+export default Home;
